+++ common/src/main/java/net/bettercombat/logic/WeaponRegistry.java
@@ -76,6 +76,7 @@ public class WeaponRegistry {
 
     private static void loadContainers(ResourceManager resourceManager) {
         Map<Identifier, AttributesContainer> containers = new HashMap();
+        var logging = BetterCombatMod.config.weapon_registry_logging;
         // Reading all attribute files
         for (var entry : resourceManager.findResources("weapon_attributes", fileName -> fileName.getPath().endsWith(".json")).entrySet()) {
             var identifier = entry.getKey();
@@ -88,7 +89,9 @@ public class WeaponRegistry {
                         .toString().replace("weapon_attributes/", "");
                 id = id.substring(0, id.lastIndexOf('.'));
                 containers.put(Identifier.of(id), container);
-                System.out.println("Loaded container: " + id);
+                if (logging) {
+                    System.out.println("Loaded container: " + id);
+                }
             } catch (Exception e) {
                 System.err.println("Failed to parse: " + identifier);
                 e.printStackTrace();